# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Movie Web App API
  description: |
    REST API for Movie Web App.

    ## Authentication
    Using **Google OAuth2** for login and **JWT Bearer Token** for accessing endpoints that require authentication.

    ## Roles
    - **guest** — Public access without authentication
    - **user** — User that has logged in via Google OAuth
    - **admin** — Administrator with full access
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://movie-ku-api.vercel.app
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from Google OAuth login
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
      required: [success, message]
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        error:
          type: string
      required: [success, message]

    # --- Auth ---
...
    SyncMoviesRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [full, changes]
          default: full
          description: |
            Sync mode. `full` syncs all available pages from TMDB list endpoints.
            `changes` uses TMDB /movie/changes API (last 14 days).
        resume:
          type: boolean
          default: false
          description: Resume from last failed sync position. Only applies to `full` mode.
        max_pages:
          type: integer
          minimum: 1
          maximum: 500
          nullable: true
          description: Optional limit on the number of pages to sync from TMDB.
    SyncStartedData:
      type: object
      properties:
        message:
          type: string
          example: "Movies sync started in background"
        type:
          type: string
          enum: [movies, changes]
        started_at:
          type: string
          format: date-time
      required: [message, type, started_at]
    SyncStopResponseData:
      type: object
      properties:
        message:
          type: string
          example: "Stop requested. Sync will halt after the current batch."
      required: [message]

    # --- Sync ---
    SyncLog:
      type: object
      properties:
        id:
          type: string
          format: uuid
        last_sync_at:
          type: string
          format: date-time
        total_inserted:
          type: integer
        total_updated:
          type: integer
        status:
          type: string
          enum: [success, failed, stopped]
        stop_requested:
          type: boolean
          description: Whether a stop was requested before the sync finished
        created_at:
          type: string
          format: date-time
      required: [id, last_sync_at, status, created_at]

    # --- Pagination ---
...
paths:
...
  /api/admin/tmdb/sync/movies:
    post:
      tags: [Admin Sync]
      summary: Trigger movie sync from TMDB
      operationId: triggerMoviesSync
      description: |
        Start a background sync with TMDB API. Supports two modes:
        - `full` (default): Sync all movies from popular + now_playing endpoints
        - `changes`: Incremental sync — only movies changed in the last 14 days

        The sync runs asynchronously. Use GET /sync/status to monitor progress.
        Returns 409 if a sync is already running.
      security:
        - BearerAuth: []
      requestBody:
        description: Sync options
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncMoviesRequest"
      responses:
        "202":
          description: Sync started in background
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: object
                        $ref: "#/components/schemas/SyncStartedData"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: A sync is already running
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /api/admin/tmdb/sync/stop:
    post:
      tags: [Admin Sync]
      summary: Stop a running movie sync
      operationId: stopMovieSync
      description: |
        Gracefully halts a running sync after the current batch finishes.
        The sync log will be saved with a `stopped` status.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Sync stop requested
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SyncStopResponseData"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/tmdb/sync/status:
    get:
      tags: [Admin Sync]
      summary: Get live sync status
      operationId: getSyncStatus
      description: |
        Returns the current sync status. If a sync is running, returns live
        in-memory progress (processed count, current endpoint/page). Otherwise,
        returns the most recent SyncLog from the database.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Current sync status
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SyncLog"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/admin/tmdb/sync/last-sync:
    get:
      tags: [Admin Sync]
      summary: Get last completed sync log
      operationId: getLastSync
      description: |
        Returns the most recent completed sync log entry from the database.
        Unlike /sync/status, this always returns the DB record regardless of
        whether a sync is currently running.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Last sync log
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SyncLog"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden — admin only
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ==================== ADMIN USER MANAGEMENT ====================
...